Fixed DRB title and favicon
