+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGE LOG
 
+## [0.10.1]
+### Fixed 
+- Fixed favicon
+- Fixed page title for search result and read online
+
 ## [0.10.0]
 
 - Refactors app to use NextJS, NYPL Design System, Typescript
