+++ src/constants/labels.ts
@@ -43,6 +43,7 @@ export const documentTitles = {
   search: "Search Results | Digital Research Books Beta | NYPL",
   workItem: "Item Details | Digital Research Books Beta | NYPL",
   editionItem: "Edition Details | Digital Research Books Beta | NYPL",
+  readItem: "Read Online | Digital Research Books Beta | NYPL",
 };
 
 export const yearsType = { start: "Start", end: "End" };
